Monroe Doctrine
Changes:
Wilson Path
Domician Occupation Path
Nicarauguan Occupation Path

TEASER:

The roaring twenties, a time of false prosperity for the United States as the rest of the world attempted to recover from the traumatic events that had untold the past decade, yet it wasn’t lost on the American people. They simply wanted a change in direction from what they had been with in the past, be the change in the United States, don't’ stay cool with Coolidge, have a terminally deprived presidency or even expand the military budget contrary to what your voters want.