Minor Error Fixing
Please Review my edits
(mostly just indent fixing)
I have no intention to contributing to this mod, I am simply using this as an excuse to scrape of the rust. I only plan on error fixing.
I wrote down a before and after test and it seem hoi4 is saying that I at least removed around 2k errors, Take that with a grain of salt it's hoi4.
Enjoy